@@ -1,26 +1,32 @@
 package actions
 
 import (
-	"io/ioutil"
+	"io"
+	"io/fs"
 	"log"
 	"strings"
 
 	"github.com/gin-gonic/gin"
-	"github.com/markbates/pkger"
 )
 
-// StaticHandler handles Static assets
-func StaticHandler(c *gin.Context) {
+func newStaticHandler(assets fs.FS) func(c *gin.Context) {
+	return func(c *gin.Context) {
+		staticHandler(c, assets)
+	}
+}
+
+// staticHandler handles Static assets
+func staticHandler(c *gin.Context, assets fs.FS) {
 	p := c.Param("filepath")
 	log.Println("/assets" + p)
-	f, err := pkger.Open("/assets" + p)
+	f, err := assets.Open("assets" + p)
 	if err != nil {
 		c.AbortWithError(404, err)
 		return
 	}
 	defer f.Close()
 
-	body, err := ioutil.ReadAll(f)
+	body, err := io.ReadAll(f)
 	if err != nil {
 		c.AbortWithError(400, err)
 		return
